requirements.txt
#albumentations==1.3.0
#einops==0.4.1
#hydra-core==1.2.0
#pytorch>=1.12.0
#pytorch-lightning==1.7.7
#scikit-learn>=1.0.2
#wandb==0.13.4

# Server requirements
flask==3.0.0
python-dotenv==1.0.0